+++ Sources/Sentry/SentryUIViewControllerSwizziling.m
@@ -3,6 +3,9 @@
 #import "SentryPerformanceTracker.h"
 #import "SentrySwizzle.h"
 #import "SentryUIViewControllerPerformanceTracker.h"
+#import <SentryInAppLogic.h>
+#import <SentryOptions.h>
+#import <UIViewController+Sentry.h>
 #import <objc/runtime.h>
 
 #if SENTRY_HAS_UIKIT
@@ -10,8 +13,12 @@
 
 @implementation SentryUIViewControllerSwizziling
 
-+ (void)start
+static SentryInAppLogic *inAppLogic;
+
++ (void)startWithOptions:(SentryOptions *)options
 {
+    inAppLogic = [[SentryInAppLogic alloc] initWithInAppIncludes:options.inAppIncludes
+                                                   inAppExcludes:options.inAppExcludes];
     [SentryUIViewControllerSwizziling swizzleViewControllerInits];
 }
 
@@ -43,20 +50,82 @@
             return SentrySWCallOriginal(nibName, bundle);
         }),
         SentrySwizzleModeOncePerClassAndSuperclasses, swizzleViewControllerInitWithNib);
+
+    [SentryUIViewControllerSwizziling swizzleRootViewController];
+}
+
+/**
+ * If an app targets, for example, iOS 13 or lower, the UIKit inits the initial/root view controller
+ * before the SentrySDK is initialized. Therefore, we manually call swizzle here not to lose
+ * auto-generated transactions for the initial view controller. As we use
+ * SentrySwizzleModeOncePerClassAndSuperclasses, we don't have to worry about swizzling twice. We
+ * could also use objc_getClassList to lookup sub classes of UIViewController, but the lookup can
+ * take around 60ms, which is not acceptable.
+ */
++ (void)swizzleRootViewController
+{
+    if (![UIApplication respondsToSelector:@selector(sharedApplication)]) {
+        NSString *message = @"UIViewControllerSwizziling: UIApplication doesn't respont to "
+                            @"sharedApplication. Skipping swizzleRootViewController.";
+        [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+        return;
+    }
+
+    UIApplication *app = [UIApplication performSelector:@selector(sharedApplication)];
+
+    if (app == nil) {
+        NSString *message = @"UIViewControllerSwizziling: UIApplication is nil. Skipping "
+                            @"swizzleRootViewController.";
+        [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+        return;
+    }
+
+    if (app.delegate == nil) {
+        NSString *message = @"UIViewControllerSwizziling: UIApplicationDelegate is nil. Skipping "
+                            @"swizzleRootViewController.";
+        [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+        return;
+    }
+
+    // Check if delegate responds to window, which it doesn't have to.
+    if (![app.delegate respondsToSelector:@selector(window)]) {
+        NSString *message
+            = @"UIApplicationDelegate.window is nil. Skipping swizzleRootViewController.";
+        [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+        return;
+    }
+
+    if (app.delegate.window == nil) {
+        NSString *message = @"UIViewControllerSwizziling UIApplicationDelegate.window is nil. "
+                            @"Skipping swizzleRootViewController.";
+        [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+        return;
+    }
+
+    UIViewController *rootViewController = app.delegate.window.rootViewController;
+    if (rootViewController == nil) {
+        NSString *message
+            = @"UIViewControllerSwizziling UIApplicationDelegate.window.rootViewController is nil. "
+              @"Skipping swizzleRootViewController.";
+        [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+        return;
+    }
+
+    NSArray<UIViewController *> *allViewControllers = rootViewController.descendantViewControllers;
+
+    for (UIViewController *viewController in allViewControllers) {
+        Class viewControllerClass = [viewController class];
+        if (viewControllerClass != nil) {
+            NSString *message = @"UIViewControllerSwizziling Calling swizzleRootViewController.";
+            [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+            [SentryUIViewControllerSwizziling swizzleViewControllerSubClass:viewControllerClass];
+        }
+    }
 }
 
 + (void)swizzleViewControllerSubClass:(Class)class
 {
-    // Swizzling only classes from the user app module to avoid track every UIKit view controller
-    // interaction.
-    static const char *appImage = nil;
-    if (appImage == nil) {
-        if ([UIApplication respondsToSelector:@selector(sharedApplication)]) {
-            UIApplication *app = [UIApplication performSelector:@selector(sharedApplication)];
-            appImage = class_getImageName(app.delegate.class);
-        }
-    }
-    if (appImage == nil || strcmp(appImage, class_getImageName(class)) != 0)
+    if (![SentryUIViewControllerSwizziling shouldSwizzleViewController:class])
         return;
 
     // This are the five main functions related to UI creation in a view controller.
@@ -66,6 +135,18 @@
     [SentryUIViewControllerSwizziling swizzleViewDidLoad:class];
     [SentryUIViewControllerSwizziling swizzleViewWillAppear:class];
     [SentryUIViewControllerSwizziling swizzleViewDidAppear:class];
+}
+
+/**
+ * For testing.
+ */
++ (BOOL)shouldSwizzleViewController:(Class)class
+{
+    // Swizzling only inApp classes to avoid track every UIKit view controller
+    // interaction.
+    NSString *classImageName = [NSString stringWithCString:class_getImageName(class)
+                                                  encoding:NSUTF8StringEncoding];
+    return [inAppLogic isInApp:classImageName];
 }
 
 + (void)swizzleLoadView:(Class)class