+++ Tests/SentryTests/SentrySpanTests.swift
@@ -137,12 +137,49 @@ class SentrySpanTests: XCTestCase {
         let serialization = span.serialize()
         XCTAssertEqual(serialization["span_id"] as? String, span.context.spanId.sentrySpanIdString)
         XCTAssertEqual(serialization["trace_id"] as? String, span.context.traceId.sentryIdString)
-        XCTAssertEqual(serialization["timestamp"] as? String, TestData.timestampAs8601String)
-        XCTAssertEqual(serialization["start_timestamp"] as? String, TestData.timestampAs8601String)
+        XCTAssertEqual(serialization["timestamp"] as? TimeInterval, TestData.timestamp.timeIntervalSince1970)
+        XCTAssertEqual(serialization["start_timestamp"] as? TimeInterval, TestData.timestamp.timeIntervalSince1970)
         XCTAssertEqual(serialization["type"] as? String, SpanContext.type)
         XCTAssertEqual(serialization["sampled"] as? String, "false")
         XCTAssertNotNil(serialization["data"])
         XCTAssertEqual((serialization["data"] as! Dictionary)[fixture.extraKey], fixture.extraValue)
+    }
+    
+    func testSerialization_WithNoData() {
+        let span = fixture.getSut()
+        
+        let serialization = span.serialize()
+        XCTAssertNil(serialization["data"])
+    }
+    
+    func testTraceHeaderNotSampled() {
+        let span = fixture.getSut()
+        let header = span.toTraceHeader()
+        
+        XCTAssertEqual(header.traceId, span.context.traceId)
+        XCTAssertEqual(header.spanId, span.context.spanId)
+        XCTAssertEqual(header.sampled, .no)
+        XCTAssertEqual(header.value(), "\(span.context.traceId)-\(span.context.spanId)-0")
+    }
+    
+    func testTraceHeaderSampled() {
+        let span = SentrySpan(context: SpanContext(operation: fixture.someOperation, sampled: .yes))
+        let header = span.toTraceHeader()
+        
+        XCTAssertEqual(header.traceId, span.context.traceId)
+        XCTAssertEqual(header.spanId, span.context.spanId)
+        XCTAssertEqual(header.sampled, .yes)
+        XCTAssertEqual(header.value(), "\(span.context.traceId)-\(span.context.spanId)-1")
+    }
+    
+    func testTraceHeaderUndecided() {
+        let span = SentrySpan(context: SpanContext(operation: fixture.someOperation, sampled: .undecided))
+        let header = span.toTraceHeader()
+        
+        XCTAssertEqual(header.traceId, span.context.traceId)
+        XCTAssertEqual(header.spanId, span.context.spanId)
+        XCTAssertEqual(header.sampled, .undecided)
+        XCTAssertEqual(header.value(), "\(span.context.traceId)-\(span.context.spanId)")
     }
     
     @available(tvOS 10.0, *)